# Server/Twitter-swagger.yaml
openapi: 3.0.3
info:
  title: Twitter API
  description: ''
  contact:
    email: apiteam@swagger.io
  version: 1.0.11
servers:
  - url: http://localhost:4000
...
paths:
...
  /users/change-password:
    post:
      tags:
        - user
      summary: Change password
      description: Change password
      operationId: ChangePassword
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  example: 123123
                password:
                  type: string
                  example: 123123123
                confirm_password:
                  type: string
                  example: 123123123
        required: true
      responses:
        default:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Change password successfully
    ####### Get my profile #######
  /users/me:
    get:
      tags:
        - user
      summary: Get my profile
      description: ''
      operationId: getMe
      security:
        - BearerAuth: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              type: object
              properties:
                bio:
                  type: string
                  example: 'This is Bio'
                name:
                  type: string
                  example: 'Thanh002'
                location:
                  type: string
                  example: 'TPHCM'
                username:
                  type: string
                  example: 'Tonionn2112'
      responses:
        default:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Update me successfully
                $ref: '#/components/schemas/UserProfile'
    ####### Update my profile #######
    patch:
      tags:
        - user
      summary: Update my profile
      description: ''
      operationId: updateMe
      security:
        - BearerAuth: []
      responses:
        default:
          description: Get my successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Update my profile successfully
                $ref: '#/components/schemas/UserProfile'
      ####### Get user profile #######
...
components:
  schemas:
...
    UserProfile:
      type: 'object'
      properties:
        _id:
          type: 'string'
          format: MongoId
          example: '65533f87da126e68ddc52e8a'
        email:
          type: 'string'
          example: 'thanh038@gmail.com'
        name:
          type: 'string'
          example: 'Thanh1234'
        date_of_birth:
          type: 'string'
          format: ISO8601
          example: 2023-09-27T15:51:28.772Z
        created_at:
          type: 'string'
          format: ISO8601
          example: 2023-11-14T09:36:07.538Z
        updated_at:
          type: 'string'
          format: ISO8601
          example: 2023-11-14T09:36:07.538Z
        verify:
          $ref: '#/components/schemas/UserVerify'
        bio:
          type: 'string'
          example: ''
        location:
          type: 'string'
          example: ''
        website:
          type: 'string'
          example: ''
        username:
          type: 'string'
          example: 'user65533f87da126e68ddc52e8a'
        avatar:
          type: 'string'
          example: ''
        cover_photo:
          type: 'string'
          example: ''
    UserVerify:
      type: number
      enum: [Unverified, Verified, Banned]
      example: 1
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT